Fix Terraform bootstrap script
I had a suspicion out of nowhere that this file would cause Terraform deployments to pull from the original repo and not this fork. It turns out I was right. Also if anyone ever reads this, a low-hanging-fruit commit to the original repo is fixing this file to pull from the new org name `telekom-security` instead of `dtag-dev-sec`