Shrink Build Timeout to 1h
Last build was finally successful, and took 00:55:59 with 32 CPU cores. This is a little under half of the timeout prior to this commit. This has been a long, weird journey and I'm glad it's over. 

I still can't get over the fact that building all 29 containers for T-Pot on a 32 CPU system takes about an hour. Roughly speaking that's about 1 container every two minutes which is actually pretty low if you think about it.